+++ first-demo/.godot/mono/temp/obj/Debug/first_demo.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("first_demo")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+427352845e0d24f642085c0af7535e2707460ed0")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+3b54951b0f634f1b8e8b182e9cbe74eade9b37d6")]
 [assembly: System.Reflection.AssemblyProductAttribute("first_demo")]
 [assembly: System.Reflection.AssemblyTitleAttribute("first_demo")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]